Address PR Comments
