+++ Elements/include/gambit/Elements/module_macros_incore_defs.hpp
@@ -67,7 +67,6 @@
 #include "gambit/Elements/module_macros_common.hpp"
 #include "gambit/Elements/safety_bucket.hpp"
 #include "gambit/Elements/ini_functions.hpp"
-#include "gambit/Elements/elements_extras.hpp"
 #include "gambit/Elements/terminator.hpp"
 #include "gambit/Utils/static_members.hpp"
 #include "gambit/Utils/exceptions.hpp"
