+++ Elements/include/gambit/Elements/module_macros_inmodule_defs.hpp
@@ -40,7 +40,6 @@
 #include "gambit/Elements/safety_bucket.hpp"
 #include "gambit/Elements/module_macros_common.hpp"
 #include "gambit/Utils/exceptions.hpp"
-#include "gambit/Elements/elements_extras.hpp"
 #include "gambit/Utils/util_macros.hpp"
 #include "gambit/Models/safe_param_map.hpp"
