+++ Elements/include/gambit/Elements/standalone_module.hpp
@@ -27,6 +27,8 @@
 #include "gambit/Utils/static_members.hpp"
 #include "gambit/Utils/stream_overloads.hpp"
 #include "gambit/Elements/module_macros_incore.hpp"
+#include "gambit/Printers/baseprinter.hpp"
+#include "gambit/Printers/printermanager.hpp"
 
 using namespace Gambit;
 using std::cout;
