+++ Printers/src/printermanager.cpp
@@ -290,6 +290,7 @@ namespace Gambit
     }
 
   }
+
 }
 
 #undef DBUG
